# internal/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Upload Assistant API",
    "version": "2.0.0-dev",
    "description": "Local-first Chinese PT retorrent automation with durable steps, rule gates, and auditable evidence."
  },
  "servers": [{"url": "http://127.0.0.1:8080"}],
  "security": [{"bearerAuth": []}],
  "paths": {
...
    "/api/v2/sites/{site_code}": {
      "put": {
        "operationId": "putSite",
        "description": "Create or update site display metadata, aliases, tags, and enabled state. Aliases never replace the canonical site code in workflow requests; a configured site does not imply runtime adapter support.",
        "parameters": [{"$ref": "#/components/parameters/SiteCode"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": false, "required": ["name", "adapter", "enabled", "aliases", "tags"], "properties": {"name": {"type": "string", "minLength": 1, "maxLength": 100}, "adapter": {"type": "string", "enum": ["config_only", "nexusphp", "mteam_api", "ttg"]}, "enabled": {"type": "boolean"}, "aliases": {"type": "array", "items": {"type": "string", "maxLength": 64}}, "tags": {"type": "array", "items": {"type": "string", "maxLength": 32}}}}}}},
        "responses": {"200": {"description": "Site metadata saved and audited", "content": {"application/json": {"schema": {"type": "object"}}}}, "409": {"$ref": "#/components/responses/Problem"}}
      }
    },
...
    "/api/v2/sites/{site_code}/rules": {
      "get": {"operationId": "listSiteRuleRevisions", "parameters": [{"$ref": "#/components/parameters/SiteCode"}], "responses": {"200": {"description": "Rule revision history", "content": {"application/json": {"schema": {"type": "object"}}}}}}
    },
...
    "/api/v2/sites/{site_code}/rules/active": {
      "get": {"operationId": "getActiveSiteRules", "parameters": [{"$ref": "#/components/parameters/SiteCode"}], "responses": {"200": {"description": "Active approved rule revision", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RuleEnvelope"}}}}, "404": {"$ref": "#/components/responses/Problem"}}}
    },
    "/api/v2/site-rules/import": {
      "post": {
        "operationId": "importSiteRuleRevision",
        "description": "Imports structured Markdown as an immutable draft. A draft with source.complete=false cannot be approved or activated.",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": false, "required": ["markdown"], "properties": {"markdown": {"type": "string", "minLength": 1, "description": "YAML front matter followed by at most 8 MiB of decoded UTF-8 rule text; keep source.complete=false until the relevant current rules have been fully reviewed."}}}}}},
        "responses": {"201": {"description": "Immutable draft imported", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RuleEnvelope"}}}}, "400": {"$ref": "#/components/responses/Problem"}}
      }
    },
...
    "/api/v2/site-rules/{revision_id}": {
      "get": {"operationId": "getSiteRuleRevision", "parameters": [{"$ref": "#/components/parameters/RevisionID"}], "responses": {"200": {"description": "Rule revision", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RuleEnvelope"}}}}, "404": {"$ref": "#/components/responses/Problem"}}}
    },
    "/api/v2/site-rules/{revision_id}/markdown": {
      "get": {"operationId": "getSiteRuleMarkdown", "parameters": [{"$ref": "#/components/parameters/RevisionID"}], "responses": {"200": {"description": "Checksum-verified Markdown", "content": {"text/markdown": {"schema": {"type": "string"}}}}}}
    },
    "/api/v2/site-rules/{revision_id}/review": {
      "get": {"operationId": "getSiteRuleReview", "description": "Read the three executable hard-gate review sections (upload, download, category-aware naming), localized facts, source conflicts, advisories, persisted decisions, blockers, and progress for the exact revision fingerprint.", "parameters": [{"$ref": "#/components/parameters/RevisionID"}], "responses": {"200": {"description": "Rule review workspace", "content": {"application/json": {"schema": {"type": "object"}}}}, "404": {"$ref": "#/components/responses/Problem"}}}
    },
    "/api/v2/site-rules/{revision_id}/review/{section}": {
      "put": {
        "operationId": "putSiteRuleReviewSection",
        "description": "Persist one human section decision bound to the exact current fingerprint. A needs_changes decision requires a comment and blocks approval.",
        "parameters": [{"$ref": "#/components/parameters/RevisionID"}, {"name": "section", "in": "path", "required": true, "schema": {"type": "string", "enum": ["upload_limit", "download_limit", "naming"]}}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": false, "required": ["fingerprint", "decision"], "properties": {"fingerprint": {"type": "string", "pattern": "^[a-fA-F0-9]{64}$"}, "decision": {"type": "string", "enum": ["confirmed", "needs_changes"]}, "comment": {"type": "string"}}}}}},
        "responses": {"200": {"description": "Section decision saved and audited", "content": {"application/json": {"schema": {"type": "object"}}}}, "409": {"$ref": "#/components/responses/Problem"}}
      }
    },
    "/api/v2/site-rules/{revision_id}/corrections/{section}": {
      "post": {
        "operationId": "correctSiteRuleHardGate",
        "description": "Derive a new immutable draft with explicitly corrected upload-limit, download-limit, or category naming data. Review comments are audit-only and are never interpreted as executable configuration. The source revision remains unchanged, section confirmations reset for the new fingerprint, and the result is neither approved nor activated.",
        "parameters": [{"$ref": "#/components/parameters/RevisionID"}, {"name": "section", "in": "path", "required": true, "schema": {"type": "string", "enum": ["upload_limit", "download_limit", "naming"]}}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": false, "required": ["fingerprint", "data", "comment"], "properties": {"fingerprint": {"type": "string", "pattern": "^[a-fA-F0-9]{64}$"}, "data": {"type": "object", "description": "Section-specific corrected executable data. upload_limit accepts upload and seedbox_upload; naming accepts the complete naming object."}, "comment": {"type": "string", "minLength": 1, "maxLength": 4000}}}}}},
        "responses": {"201": {"description": "Corrected immutable draft created or reused by fingerprint", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RuleEnvelope"}}}}, "400": {"$ref": "#/components/responses/Problem"}, "409": {"$ref": "#/components/responses/Problem"}}
      }
    },
    "/api/v2/site-rules/{revision_id}/approve": {
      "post": {
        "operationId": "approveSiteRuleRevision",
        "parameters": [{"$ref": "#/components/parameters/RevisionID"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": false, "required": ["fingerprint"], "properties": {"fingerprint": {"type": "string", "pattern": "^[a-fA-F0-9]{64}$"}, "comment": {"type": "string"}}}}}},
        "responses": {"200": {"description": "Complete rule revision approved after all three executable hard gates were confirmed for the exact fingerprint", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RuleEnvelope"}}}}, "409": {"$ref": "#/components/responses/Problem"}}
      }
    },
    "/api/v2/site-rules/{revision_id}/activate": {
      "post": {"operationId": "activateSiteRuleRevision", "parameters": [{"$ref": "#/components/parameters/RevisionID"}], "responses": {"200": {"description": "Approved revision activated", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RuleEnvelope"}}}}, "409": {"$ref": "#/components/responses/Problem"}}}
    },
    "/api/v2/site-rules/{revision_id}/discard": {
      "post": {
        "operationId": "discardSiteRuleDraft",
        "description": "Remove one pending draft from the maintenance workspace by marking it retired. Immutable source, fingerprint, review evidence, and audit history are retained.",
        "parameters": [{"$ref": "#/components/parameters/RevisionID"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": false, "required": ["fingerprint", "confirm"], "properties": {"fingerprint": {"type": "string", "pattern": "^[a-fA-F0-9]{64}$"}, "confirm": {"type": "boolean", "const": true}}}}}},
        "responses": {"200": {"description": "Draft retired and retained for audit", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RuleEnvelope"}}}}, "400": {"$ref": "#/components/responses/Problem"}, "409": {"$ref": "#/components/responses/Problem"}}
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "ua_API_token"}
    },
    "parameters": {
...
      "RevisionID": {"name": "revision_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
      "SiteCode": {"name": "site_code", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^[A-Z0-9_-]+$"}},
...
    },
    "responses": {
      "Problem": {"description": "Structured API problem", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Problem"}}}}
    },
    "schemas": {
...
      "Blocker": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {"code": {"type": "string"}, "message": {"type": "string"}, "site_code": {"type": "string"}, "obligation_id": {"type": "string"}},
        "additionalProperties": true
      },
      "NextAction": {
        "type": "object",
        "required": ["action"],
        "properties": {"action": {"type": "string"}, "description": {"type": "string"}, "parameters": {"type": "object", "additionalProperties": true}},
        "additionalProperties": true
      },
...
      "RuleEnvelope": {"type": "object", "required": ["ok", "status", "site_code", "rule_revision_id", "fingerprint", "revision", "blockers", "next_actions"], "properties": {"ok": {"type": "boolean"}, "status": {"type": "string", "enum": ["draft", "approved", "retired"]}, "site_code": {"type": "string"}, "rule_revision_id": {"type": "string", "format": "uuid"}, "fingerprint": {"type": "string"}, "revision": {"type": "object"}, "blockers": {"type": "array", "items": {"$ref": "#/components/schemas/Blocker"}}, "next_actions": {"type": "array", "items": {"$ref": "#/components/schemas/NextAction"}}}},
...
      "Problem": {"type": "object", "required": ["ok", "status", "error", "blockers", "next_actions"], "properties": {"ok": {"type": "boolean", "const": false}, "status": {"type": "string", "const": "failed"}, "error": {"type": "object", "required": ["code", "detail"], "properties": {"code": {"type": "string"}, "detail": {"type": "string"}}}, "blockers": {"type": "array", "items": {"$ref": "#/components/schemas/Blocker"}}, "next_actions": {"type": "array", "items": {"$ref": "#/components/schemas/NextAction"}}}}
    }
  }
}